@@ -10,11 +10,12 @@ dependencies:
 dev_dependencies:
   flutter_driver:
     sdk: flutter
-  integration_test: ^1.0.2
+  integration_test:
+    sdk: flutter
   pedantic: ^1.9.2
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0-0'
