@@ -5,7 +5,7 @@ repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 version: 0.1.0
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-0"
   flutter: ">=1.20.0"
 
 dependencies:
