@@ -8,7 +8,7 @@ dependencies:
   flutter:
     sdk: flutter
   meta: ^1.1.8
-  plugin_platform_interface: ^1.0.2
+  plugin_platform_interface: ^1.1.0-nullsafety.2
 
 dev_dependencies:
   flutter_test:
@@ -17,5 +17,5 @@ dev_dependencies:
   pedantic: ^1.9.2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0"
   flutter: ">=1.9.1+hotfix.4"
