@@ -11,9 +11,9 @@ import 'package:integration_test/integration_test.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  IosDeviceInfo iosInfo;
-  AndroidDeviceInfo androidInfo;
-  WebBrowserInfo webBrowserInfo;
+  IosDeviceInfo? iosInfo;
+  AndroidDeviceInfo? androidInfo;
+  WebBrowserInfo? webBrowserInfo;
 
   setUpAll(() async {
     final deviceInfoPlugin = DeviceInfoPlugin();
@@ -30,13 +30,13 @@ void main() {
 
   testWidgets('Can get non-null device model', (WidgetTester tester) async {
     if (Platform.isIOS) {
-      expect(iosInfo.model, isNotNull);
+      expect(iosInfo?.model, isNotNull);
     } else if (Platform.isAndroid) {
-      expect(androidInfo.model, isNotNull);
+      expect(androidInfo?.model, isNotNull);
     }
 
     if (kIsWeb) {
-      expect(webBrowserInfo.userAgent, isNotNull);
+      expect(webBrowserInfo?.userAgent, isNotNull);
     }
   });
 }
