@@ -34,20 +34,23 @@ class DeviceInfoPlugin {
   }
 
   static bool _disablePlatformOverride = false;
-  static DeviceInfoPlatform __platform;
+  static DeviceInfoPlatform? __platform;
 
   // This is to manually endorse the Linux plugin until automatic registration
   // of dart plugins is implemented.
   // See https://github.com/flutter/flutter/issues/52267 for more details.
   static DeviceInfoPlatform get _platform {
-    __platform ??= !kIsWeb && Platform.isLinux && !_disablePlatformOverride
+    var platform = __platform;
+
+    platform ??= !kIsWeb && Platform.isLinux && !_disablePlatformOverride
         ? DeviceInfoLinux()
         : DeviceInfoPlatform.instance;
-    return __platform;
+
+    return __platform = platform;
   }
 
   /// This information does not change from call to call. Cache it.
-  AndroidDeviceInfo _cachedAndroidDeviceInfo;
+  AndroidDeviceInfo? _cachedAndroidDeviceInfo;
 
   /// Information derived from `android.os.Build`.
   ///
@@ -56,7 +59,7 @@ class DeviceInfoPlugin {
       _cachedAndroidDeviceInfo ??= await _platform.androidInfo();
 
   /// This information does not change from call to call. Cache it.
-  IosDeviceInfo _cachedIosDeviceInfo;
+  IosDeviceInfo? _cachedIosDeviceInfo;
 
   /// Information derived from `UIDevice`.
   ///
@@ -65,7 +68,7 @@ class DeviceInfoPlugin {
       _cachedIosDeviceInfo ??= await _platform.iosInfo();
 
   /// This information does not change from call to call. Cache it.
-  LinuxDeviceInfo _cachedLinuxDeviceInfo;
+  LinuxDeviceInfo? _cachedLinuxDeviceInfo;
 
   /// Information derived from `/etc/os-release`.
   ///
@@ -74,14 +77,14 @@ class DeviceInfoPlugin {
       _cachedLinuxDeviceInfo ??= await _platform.linuxInfo();
 
   /// This information does not change from call to call. Cache it.
-  WebBrowserInfo _cachedWebBrowserInfo;
+  WebBrowserInfo? _cachedWebBrowserInfo;
 
   /// Information derived from `Navigator`.
   Future<WebBrowserInfo> get webBrowserInfo async =>
       _cachedWebBrowserInfo ??= await _platform.webBrowserInfo();
 
   /// This information does not change from call to call. Cache it.
-  MacOsDeviceInfo _cachedMacosDeviceInfo;
+  MacOsDeviceInfo? _cachedMacosDeviceInfo;
 
   /// Returns device information for macos. Information sourced from Sysctl.
   Future<MacOsDeviceInfo> get macOsInfo async =>
