@@ -4,4 +4,6 @@
 
 #include "generated_plugin_registrant.h"
 
-void fl_register_plugins(FlPluginRegistry* registry) {}
+
+void fl_register_plugins(FlPluginRegistry* registry) {
+}
